refactorisation du code
