+++ src/pc/units/forceUnits.ts
@@ -1,14 +1,10 @@
-import { AlgebraicNode } from "#root/tree/nodes/algebraicNode";
-import { VariableNode } from "#root/tree/nodes/variables/variableNode";
-import exp from "constants";
-import { Measure } from "../measure/measure";
 import { Unit } from "./unit";
-import { BasicUnit } from "./basicUnit";
+import { BaseTenUnit } from "./baseTenUnit";
 
 export type forceValues = "kN" | "hN" | "daN" | "N" | "dN" | "cN" | "mN";
 const forces = ["kN", "hN", "daN", "N", "dN", "cN", "mN"];
 
-export class ForceUnit extends BasicUnit {
+export class ForceUnit extends BaseTenUnit {
   static readonly kN = new ForceUnit("kN");
   static readonly hN = new ForceUnit("hN");
   static readonly daN = new ForceUnit("daN");
